Refactor AccessRights inner class out of Agent class
